Moved menu ui activities to menus_ui package
