+++ app/src/main/java/de/ironjan/mensaupb/menus_ui/MenuListingFragment.java
@@ -20,7 +20,6 @@ import org.slf4j.LoggerFactory;
 
 import de.ironjan.mensaupb.BuildConfig;
 import de.ironjan.mensaupb.R;
-import de.ironjan.mensaupb.activities.MenusNavigationCallback;
 import de.ironjan.mensaupb.prefs.InternalKeyValueStore_;
 import de.ironjan.mensaupb.stw.opening_times.OpeningTimesKeeper;
 import de.ironjan.mensaupb.sync.AccountCreator;
