+++ app/src/main/java/de/ironjan/mensaupb/menus_ui/MenusNavigationCallback.java
@@ -1,4 +1,4 @@
-package de.ironjan.mensaupb.activities;
+package de.ironjan.mensaupb.menus_ui;
 
 /**
  * Interface to handle navigation from a menu listing
